Allow ALL of the Qt Color roles to be accessed
This is to work around TooltipBase not being set properly when reading
in colors from GTK+. Furthermore, this change allows qtnotifydaemon to
be more self-documenting, since the names here are the same ones used by
Qt internally (just prefixed with a Qt to make it more obvious). The
names are case insensitive so the user doesn't have to
WorryAboutCamelCasing.